Move view-related model methods into decorators
