+++ Gemfile
@@ -13,6 +13,7 @@ gem 'sassc-rails' # Use SASS for stylesheets
 gem 'slim-rails' # Use Slim for views
 gem 'uglifier', '>= 1.3.0' # Use Uglifier as compressor for JavaScript assets
 gem 'jbuilder', '~> 2.5' # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'active_decorator' # Separate view code while keeping it attached to the model
 
 ### Javascript
 gem 'jquery-rails' # Add jQuery
