+++ app/decorators/local_area_decorator.rb
@@ -1,0 +1,7 @@
+module LocalAreaDecorator
+
+  def label
+    name
+  end
+
+end
