+++ app/decorators/manager_decorator.rb
@@ -1,0 +1,7 @@
+module ManagerDecorator
+
+  def label
+    name
+  end
+
+end
