+++ app/decorators/registration_decorator.rb
@@ -1,0 +1,7 @@
+module RegistrationDecorator
+
+  def label
+    name
+  end
+
+end
