+++ app/models/application_record.rb
@@ -5,10 +5,6 @@ class ApplicationRecord < ActiveRecord::Base
   include Searchable
   include Parentable
 
-  def label
-    name
-  end
-
   def has_region_association?
     respond_to?(:countries) || respond_to?(:provinces) || respond_to?(:local_areas)
   end
