+++ app/models/province.rb
@@ -1,31 +1,26 @@
 class Province < ApplicationRecord
 
+  # Extensios
   include Manageable
+  searchable_columns %w[province_code country_code]
 
+  # Associations
   belongs_to :country, foreign_key: :country_code, primary_key: :country_code
   has_many :local_areas, inverse_of: :province, foreign_key: :province_code, primary_key: :province_code, dependent: :delete_all
 
   has_many :venues, foreign_key: :province, primary_key: :province_code
   has_many :events, through: :venues
   
-  searchable_columns %w[province_code country_code]
-  alias_method :parent, :country
-
-  default_scope { order(province_code: :desc) }
-
+  # Validations
   validates_presence_of :province_code, :country_code
 
-  def name
-    ISO3166::Country[country_code].subdivisions[province_code]['name']
-  end
+  # Scopes
+  default_scope { order(province_code: :desc) }
 
-  def label
-    "#{name}, #{country_code}"
-  end
+  # Delegations
+  alias_method :parent, :country
 
-  def self.get_label province_code, country_code
-    "#{ISO3166::Country[country_code].subdivisions[province_code]['name']}, #{country_code}"
-  end
+  # Methods
 
   def contains? venue
     venue.province == province_code
@@ -34,10 +29,6 @@ class Province < ApplicationRecord
   def managed_by? manager, super_manager: false
     return true if managers.include?(manager) && !super_manager
     return country.managed_by?(manager)
-  end
-
-  def to_s
-    label
   end
 
 end
