+++ app/models/venue.rb
@@ -1,25 +1,29 @@
 class Venue < ApplicationRecord
 
+  # Extensions
   include Manageable
-  acts_as_mappable lat_column_name: :latitude, lng_column_name: :longitude
-  
+
   nilify_blanks
+  acts_as_mappable lat_column_name: :latitude, lng_column_name: :longitude
+  searchable_columns %w[name street city province_code country_code]
+
+  # Associations
   belongs_to :country, foreign_key: :country_code, primary_key: :country_code, optional: true
   belongs_to :province, foreign_key: :province_code, primary_key: :province_code, optional: true
   has_many :events
 
+  # Validations
   validates :street, presence: true
   validates :country_code, presence: true
   validates :latitude, :longitude, presence: true
 
-  searchable_columns %w[name street city province_code country_code]
-  alias_method :parent, :province
-
+  # Scopes
   default_scope { order(updated_at: :desc) }
 
-  def label
-    name || street
-  end
+  # Delegations
+  alias_method :parent, :province
+
+  # Methods
 
   def managed_by? manager, super_manager: false
     return true if !super_manager && managers.include?(manager)
@@ -40,24 +44,6 @@ class Venue < ApplicationRecord
     [latitude, longitude]
   end
   
-  def full_address
-    [street, city, province_code, country_code].compact.join(', ')
-  end
-
-  def address
-    {
-      street: street,
-      city: city,
-      province: province_code,
-      country: country_code,
-      postcode: postcode,
-    }
-  end
-
-  def country
-    I18nData.countries(I18n.locale)[country_code]
-  end
-
   def country_code= value
     value = value.to_s.upcase
     # Only accept country codes which are in the language list
