+++ config/initializers/mailer.rb
@@ -12,7 +12,6 @@ Rails.application.configure do
   }
 
   config.action_mailer.perform_deliveries = true
-  config.action_mailer.default_url_options = { host: 'database.sydevelopers.com' }
   config.action_mailer.default_options = { from: "SY Program Database <#{ENV['SMTP_EMAIL']}>" }
   config.action_mailer.raise_delivery_errors = true
 end
