+++ test/fixtures/venues.yml
@@ -4,10 +4,9 @@ phoenix_center:
   name: Sahaja Yoga Meditation Center
   latitude: 33.4878679
   longitude: -112.0307763
-  contact_email: azsahajayoga@gmail.com
   street: 2434 E Osborn Rd
   city: Phoenix
-  province: AZ
+  province_code: AZ
   country_code: US
   postcode: '85016'
 
@@ -15,10 +14,9 @@ chula_vista_eastlake:
   name: Eastlake Beach Club Pavilion
   latitude: 32.6534345
   longitude: -116.9805553
-  contact_email: test@test.com
   street: 2434 E Osborn Rd
   city: Phoenix
-  province: AZ
+  province_code: AZ
   country_code: US
   postcode: '85016'
 
@@ -26,10 +24,9 @@ oak_park_community:
   name: Oak Park Community Center
   latitude: 34.1813733
   longitude: -118.7743102
-  contact_email: test@test.com
   street: 1000 Kanan Road
   city: Oak Park
-  province: CA
+  province_code: CA
   country_code: US
   postcode: '91377'
 
@@ -37,10 +34,9 @@ burbank_rec_center:
   name: McCambridge Recreation Center
   latitude: 34.1927014
   longitude: -118.3230303
-  contact_email: test@test.com
   street: 1515 N Glenoaks Blvd
   city: Burbank
-  province: CA
+  province_code: CA
   country_code: US
   postcode: '91504'
 
@@ -48,10 +44,9 @@ folsom_dance_studio:
   name: Cynthia Dance Studio
   latitude: 38.6667772
   longitude: -121.1358095
-  contact_email: test@test.com
   street: 304 S Lexington Dr
   city: Folsom
-  province: CA
+  province_code: CA
   country_code: US
   postcode: '95630'
 
@@ -59,9 +54,8 @@ fremont_center:
   name: Fremont Meditation Center
   latitude: 37.5780345
   longitude: -121.9848458
-  contact_email: test@test.com
   street: 37138 Niles Blvd
   city: Fremont
-  province: CA
+  province_code: CA
   country_code: US
   postcode: '94536'
